Revert "Feature sp4 weekly"
