+++ sql/ddl.sql
@@ -14,7 +14,6 @@ create table weekly_box
 (
     id bigint generated by default as identity,
     title varchar(255),
-    clear boolean default 0,
     primary key (id)
 );
 
@@ -29,14 +28,4 @@ create table todo_elm
     foreign key (to_do_box_id) references todo_box(id) on delete cascade
 );
 
-drop table if exists weekly_elm CASCADE;
-create table weekly_elm
-(
-    id bigint NOT NULL GENERATED ALWAYS AS IDENTITY,
-    content varchar(255),
-    completed boolean default 0,
-    weekly_box_id bigint not null,
-    primary key (id),
-    foreign key (weekly_box_id) references weekly_box(id) on delete cascade
-);
 --     id bigint generated by default as identity,
