--- src/main/java/com/handey/web/controller/history/WeeklyElmController.java
@@ -1,4 +1,0 @@
-package com.handey.web.controller.history;
-
-public class WeeklyElmController {
-}
