--- src/main/java/com/handey/web/controller/history/WeeklyParam.java
@@ -1,11 +1,0 @@
-package com.handey.web.controller.history;
-
-import lombok.Data;
-
-@Data
-public class WeeklyParam {
-
-    private String title;
-
-    private boolean fixed;
-}
