+++ src/main/java/com/handey/web/controller/home/ToDoElmParam.java
@@ -4,5 +4,9 @@ import lombok.Data;
 
 @Data
 public class ToDoElmParam {
+    private String content;
 
+    private boolean completed;
+
+    private Long toDoBoxId;
 }
