--- src/main/java/com/handey/web/service/WeeklyElmService.java
@@ -1,4 +1,0 @@
-package com.handey.web.service;
-
-public class WeeklyElmService {
-}
