+++ src/main/resources/static/css/screens/history.css
@@ -35,8 +35,8 @@ body {
     position: absolute;
     top: 0;
     left: 0;
-    width: 920px;
-    height: 50px;
+    width: 1140px;
+    height: 40px;
     margin-left: 40px;
     margin-right: 40px;
     margin-top: 20px;
