+++ src/test/java/com/handey/web/service/WeeklyServiceIntegrationTest.java
@@ -1,7 +1,7 @@
 package com.handey.web.service;
 
-import com.handey.web.domain.history.WeeklyBox;
-import com.handey.web.repository.history.WeeklyRepository;
+import com.handey.web.domain.home.WeeklyBox;
+import com.handey.web.repository.home.WeeklyRepository;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
@@ -9,6 +9,7 @@ import org.springframework.test.annotation.Commit;
 import org.springframework.transaction.annotation.Transactional;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
 
 @SpringBootTest
 @Transactional
